fix ui in android
