+++ app/(tabs)/_layout.tsx
@@ -20,7 +20,7 @@ export default function TabLayout() {
   return (
     <Tab.Navigator
       screenOptions={{
-        tabBarActiveTintColor: Colors[colorScheme ?? "light"].tint,
+        tabBarActiveTintColor: Colors["light"].tint,
         headerShown: useClientOnlyValue(false, true),
       }}
     >
