+++ app/(tabs)/multipleSession.tsx
@@ -1,11 +1,13 @@
-import { View, StyleSheet } from "react-native";
+import { View, StyleSheet, Platform } from "react-native";
 import ScreenWrapper from "@/components/ScreenWrapper";
 import CustomHeader from "@/components/ui/CustomHeader";
 import SessionForm from "@/components/forms/SessionForm";
 
 export default function MultipleSession() {
   return (
-    <View style={styles.content}>
+    <View
+      style={[styles.content, Platform.OS === "android" && { paddingTop: 40 }]}
+    >
       <View style={styles.container}>
         <CustomHeader screenTitle="New session" />
         <SessionForm />
